% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btt22-package.R
\docType{package}
\name{btt22-package}
\alias{btt22}
\alias{btt22-package}
\title{btt22: Templating functions for "Building Tidy Tools 2022"}
\description{
This package is designed to help you build the "ussie" package. At its core, it calls `usethis::use_template()` to put starter files into the package you are building.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio-conf-2022/btt22}
  \item Report bugs at \url{https://github.com/rstudio-conf-2022/btt22/issues}
}

}
\author{
\strong{Maintainer}: Ian Lyttle \email{ijlyttle@me.com} (\href{https://orcid.org/0000-0001-9962-4849}{ORCID})

Authors:
\itemize{
  \item Emma Rand \email{emma.rand@york.ac.uk} (\href{https://orcid.org/0000-0002-1358-8275}{ORCID})
}

}
\keyword{internal}
