% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{btt_get}
\alias{btt_get}
\title{Get new files for a given state}
\usage{
btt_get(state)
}
\arguments{
\item{state}{\code{character}, use \code{\link[=btt_state]{btt_state()}} to see available states.}
}
\value{
\code{state}, invisibly.
}
\description{
This function add new files to
\itemize{
\item your \code{R} directory.
\item your \code{tests} directory.
}
}
\examples{
\dontrun{
btt_get("2.1.2")
}
}
