% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{btt_reset_hard}
\alias{btt_reset_hard}
\title{Reset ussie to a fixed state}
\usage{
btt_reset_hard(state)
}
\arguments{
\item{state}{\code{character}, use \code{\link[=btt_state]{btt_state()}} to see available states.}
}
\value{
\code{state}, invisibly.
}
\description{
This function completely overwrites:
\itemize{
\item your \code{R} directory.
\item your \code{tests} directory.
\item the \code{Imports} and \code{Suggests} sections of your \code{DESCRIPTION} file.
}
}
\details{
This function must be called interactively, where the current working
directory is within an R package named \strong{ussie}. This is to prevent
you from accidentially blowing away your work in a different project.
}
\examples{
\dontrun{
btt_reset_hard("2.1.1")
}
}
