% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{btt_state}
\alias{btt_state}
\title{Get available states}
\usage{
btt_state(pattern = "^[12]\\\\.")
}
\arguments{
\item{pattern}{\code{character} regular expression, used to filter
candidate-states. Default specifies name has to start with a \code{1} or \code{2},
followed by a dot (\code{.}).}
}
\value{
\code{character} vector of all available states.
}
\description{
Returns all available states of ussie.
}
\examples{
btt_state()
}
