% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{btt_get}
\alias{btt_get}
\title{Get new files for a given state}
\usage{
btt_get(state, open = rlang::is_interactive())
}
\arguments{
\item{state}{\code{character}, use \code{\link[=btt_state]{btt_state()}} to see available states.}

\item{open}{\code{logical} specifies if new files are opened for interactive
editing}
}
\value{
\code{state}, invisibly; called for side effects.
}
\description{
This function adds new files to:
\itemize{
\item your \code{R} directory
\item your \code{testthat/tests} directory
}

If \code{open} is \code{TRUE}, it will open the new files for editing.
}
\examples{
\dontrun{
btt_get("2.1.2")
}
}
